rename CLObjects module to buffers
