+++ examples/rayscattering/accelerated/exampleInfinite.py
