+++ pytissueoptics/rayscattering/opencl/buffers/surfaceCL.py
@@ -1,6 +1,6 @@
 from typing import List, NamedTuple
 
-from pytissueoptics.rayscattering.opencl.CLObjects.CLObject import *
+from pytissueoptics.rayscattering.opencl.buffers.CLObject import *
 
 
 SurfaceCLInfo = NamedTuple("SurfaceInfo", [("firstPolygonID", int), ("lastPolygonID", int),
